requirements.txt
flask==2.2.3
mysql-connector-python==8.0.32
twilio==7.17.0